List and Details Update and CRUD Working

+++ src/app/geru.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable'
 import { User } from './geru.model';
 import 'rxjs/add/operator/map'
@@ -10,7 +10,9 @@ export class GeruService {
 
   EMISSORES_URL = '../assets/expedidores.json';
   ENDPOINT = 'http://5ad4038b33667e001462443f.mockapi.io/api/v1/users';
-
+  
+  private headers = new HttpHeaders  ({"Content-Type": "application/x-www-form-urlencoded"})
+  params = new HttpParams()
 
   constructor(private http: HttpClient) { }
 
@@ -23,8 +25,14 @@ export class GeruService {
       .map(user => user.id)
   }
 
+  editUser(id:number, info:User): Observable<User>{
+    return this.http.put<User>(`${this.ENDPOINT}/${id}`, info)
+  }
+
   deleteUser(id:number): Observable<User>{
-    return this.http.delete<User>(`${this.ENDPOINT}/${id}`)
+    if(confirm("Você tem certeza?")){
+      return this.http.delete<User>(`${this.ENDPOINT}/${id}`)
+    }
   }
 
   listUsers(): Observable<User>{
